Initial* commit
Library for doing networky stuff in C++

It is for my personal use for hobby projects. Do not use this for Enterprise Solutions (TM).
Go use Python wsgi or nginx and FastCGI or apache2 and php or something if you care about bugs and security.

So, we have Socket's, TCP::Socket's, and HTTP::Request's
Most of Sockets/TCP::Sockets stolen from previous projects.

Will try implementing Websockets next. That should be fun...