Websocket Handshake
Credit to René Nyffenegger <rene.nyffenegger@adp-gmbh.ch>
for base64.cpp and base64.h

Using OpenSSL for SHA1

Apparently this magic key handshake is for security.
But not for encryption. It's to stop people using XMLRequests
in Javascript to trick the server into making a websocket
without actually connecting. Or something. Sigh.