Refactor, Fix Websocket sending/receiving
That was pretty brutal...

Spent hours on a bug... I thought use of char instead of uint8_t
was causing it but it was because I forgot to Socket::Valid virtual.
There were other things hideously wrong... my memory is fading fast...

I don't know. It works now. I hope.

wscat is netcat but using WebSockets.
This is how I found all the bugs.