Get CGI output correctly
CGI doesn't write the HTTP/1.1 status line; the webserver must do that.

Which means instead of using Socket::Cat, we have to pass the message from the client to the process,
then read the process output, parse it for a "Status" header, then send the HTTP/1.1 status line,
send the headers, then send the rest of the process output straight on.

Had some confusion dealing with blocking...