(Not working) Trying to deal with SIGCHLD
The sigchld handler is not allowed to use mutexes, because they can deadlock; if a signal handler arrives whilst the main process is in a mutex, we are in trouble.

However, we need to be able to handle both cases where Process is destroyed by the parent (out of scope/deleted by user of the library)
and cases where the child terminates naturally.

Trying to find a way around having an event loop that the user has to manually poll...

This attempt uses a thread to poll for SIGCHLD, but it doesn't work, because the main thread can call a blocking function on a file descriptor
that becomes invalid when a SIGCHLD arrives.

We could fix it if we over-rode all the virtual Socket functions in Process, and wrapped them in a mutex.
That's a lot of stuff though.

Signal handling sucks and is terrible and why the hell was it designed this way, seriously, it is the worst thing ever.