Interrupt system calls in thread owning process
We can keep the dedicated SIGCHLD thread, which cleans up the Process variable.
Then we send a SIGUSR1 to interrupt system calls only in the thread responsible for the Process.

It seems to work. httpserver is now multithreaded so we can test things properly.
Had to use syscall(2) to get thread ID and send signal to specific thread. Might be issues that I'm not aware of.

Could generalise the thread that responds to SIGUSR1 to be a member of Socket, not Process. Then Socket sets the thread id to the calling thread
before starting any system functions that may block.

This doesn't seem like a traditional way of doing it, but it is interesting.

There are a *lot* of errors being reported in valgrind at the moment.
Also if you spam requests at httpserver too fast, it seems to die (no exception is raised, no error is printed...)