Ignore SIGPIPE
If the client closes its end early, then the server gets a SIGPIPE which was terminating it.
There was also a Transport end point not connected when the socket closes.

Wrote stress tester for the server; queries the cgi test script

It still crashes after about about 100 requests for each of 10 threads in parallel.
Server can't listen/accept on the socket anymore?

Should I allow the user to specify whether or not SIGPIPE is ignored?
In case their application wants to handle it?