Add Pipe class and Stdio
This is mainly so I can tread stdio like a Socket.
Don't want to have two seperate sockets for stdin and stdout.

Socket::Select will work because it only checks for reading.
If I included writing, things might get hard.

Pipe is a Socket so it has its own file descriptor, which it borrows from the input (not the output).
I could add a virtual "GetWriteFD" or "GetReadFD" but all these virtual functions bother me.